+++ apraw/models/__init__.py
@@ -1,9 +1,9 @@
 from .helpers.apraw_base import aPRAWBase
 from .helpers.generator import ListingGenerator
 from .helpers.item_moderation import ItemModeration, PostModeration
-from .helpers.listing import Listing
 from .helpers.streamable import Streamable
 from .reddit.comment import Comment
+from .reddit.listing import Listing
 from .reddit.message import Message
 from .reddit.redditor import Redditor
 from .reddit.submission import Submission
